Create discovery package, remove unused config settings
