+++ core/discovery/discovery.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package core
+package discovery
 
 import (
 	"math/rand"
@@ -24,10 +24,20 @@ import (
 	"github.com/op/go-logging"
 )
 
-var discLogger *logging.Logger
+var logger *logging.Logger
 
 func init() {
-	discLogger = logging.MustGetLogger("core/discovery")
+	logger = logging.MustGetLogger("discovery")
+}
+
+// Discovery is the interface that consolidates bootstrap peer membership
+// selection and validating peer selection for non-validating peers
+type Discovery interface {
+	AddNode(string) bool    // Add an address to the discovery list
+	RemoveNode(string) bool // Remove an address from the discovery list
+	GetAllNodes() []string  // Return all addresses this peer maintains
+	GetRandomNode() string  // Return a random address for this peer to connect to
+	FindNode(string) bool   // Find a node in the discovery list
 }
 
 // DiscoveryImpl is an implementation of Discovery
@@ -50,7 +60,7 @@ func NewDiscoveryImpl() *DiscoveryImpl {
 func (di *DiscoveryImpl) AddNode(address string) bool {
 	di.Lock()
 	defer di.Unlock()
-	discLogger.Debugf("About to add %v to %v", address, di.seq)
+	logger.Debugf("About to add %v to %v", address, di.seq)
 	if _, ok := di.nodes[address]; !ok {
 		di.seq = append(di.seq, address)
 		di.nodes[address] = true
